Redmineのチケットナンバー　#14264
不具合状況の説明:物件と部屋番号の出力が逆になっている
修正内容の説明:物件と部屋番号が正しく表示されている

+++ SRC/libArticle.php
@@ -53,18 +53,19 @@ function subArticle()
 				<tr>
 					<th>除外</th>
 					<td><input type="checkbox" name="sDel" value="0" <?php if ($sDel == 0) print ' checked="checked"' ?> /></td>
+					<th>部屋番号</th>
+					<td><input type="text" name="sRoom" value="" size="30" /><?php print $sRoom ?></td>
+				</tr>
+				<tr>
+					<th>物件名</th>
+					<td><input type="text" name="sArticle" value="<?php print $sArticle ?>" size="50" /></td>
+					<th>キーBox番号</th>
+					<td><input type="text" name="sKeyBox" value="<?php print $sKeyBox ?>" size="30" /></td>
+				</tr>
+				<tr>
 					<th>備考</th>
 					<td><input type="text" name="sArticleNote" value="<?php print $sArticleNote ?>" size="50" /></td>
-				</tr>
-				<tr>
-					<th>部屋番号</th>
-					<td><input type="text" name="sRoom" value="<?php print $sRoom ?>" size="30" /></td>
-					<th>3Dパース</th>
-					<td><input type="text" name="sDrawing" value="<?php print $sDrawing ?>" size="30" /></td>
-				</tr>
-				<tr>
-					<th>部屋番号</th>
-					<td><input type="text" name="sRoom" value="" size="30" /><?php print $sArticle ?></td>
+
 					<th>3Dパース</th>
 					<td><input type="text" name="sDrawing" value="<?php print $sDrawing ?>" size="30" /></td>
 				</tr>
@@ -86,6 +87,7 @@ function subArticle()
 		return;
 	}
 	$sql = fnSqlArticleList(0, $sDel, $sArticle, $sRoom, $sKeyPlace, $sArticleNote, $sKeyBox, $sDrawing, $sSellCharge, $sPage, $orderBy, $orderTo);
+
 	$res = mysqli_query($conn, $sql);
 	$row = mysqli_fetch_array($res);
 
@@ -108,6 +110,7 @@ function subArticle()
 			</tr>
 			<?php
 			$sql = fnSqlArticleList(1, $sDel, $sArticle, $sRoom, $sKeyPlace, $sArticleNote, $sKeyBox, $sDrawing, $sSellCharge, $sPage, $orderBy, $orderTo);
+			var_dump($sql);
 
 			$res = mysqli_query($conn, $sql);
 			$i = 0;
@@ -124,19 +127,29 @@ function subArticle()
 				<tr>
 					<td class="list_td<?php print $i ?>"><a href="javascript:form.act.value='articleEdit';form.articleNo.value=<?php print $articleNo ?>;form.submit();"><?php print $article ?></a></td>
 					<td class="list_td<?php print $i ?>"><?php print $room ?></td>
+
 					<td class="list_td<?php print $i ?>"><?php print $keyPlace ?></td>
+
 					<td class="list_td<?php print $i ?>"><?php print $articleNote ?></td>
+
+
+
+
+
 					<td class="list_td<?php print $i ?>"><a href="javascript:form.act.value='fManager';form.sName.value='<?php print $article ?>';form.sRoom.value='<?php print $room ?>';form.submit();">表示</a></td>
 					<td class="list_td<?php print $i ?>"><?php print $keyBox ?></td>
 					<td class="list_td<?php print $i ?>"><?php print $drawing ?></td>
 					<td class="list_td<?php print $i ?>"><?php print $sellCharge ?></td>
+
 				<?php
 				$i = ($i + 1) % 2;
 			}
 				?>
+
 		</table>
 	</div>
 <?php
+	var_dump($room);
 }
 
 
@@ -148,6 +161,7 @@ function subArticle()
 function subArticleEdit()
 {
 	$conn = fnDbConnect();
+
 
 	$sDel         = $_REQUEST['sDel'];
 	$sArticle     = $_REQUEST['sArticle'];
@@ -170,15 +184,16 @@ function subArticleEdit()
 		$res = mysqli_query($conn, $sql);
 		$row = mysqli_fetch_array($res);
 
-		$article     =  $row["ARTICLE"];
-		$room        =  $row["ROOM"];
-		$keyPlace    =  $row["KEYPLACE"];
-		$address     =  $row["ADDRESS"];
-		$articleNote =  $row["ARTICLENOTE"];
-		$keyBox      =  $row["KEYBOX"];
-		$drawing     =  $row["DRAWING"];
-		$sellCharge  =  $row["SELLCHARGE"];
-		$del         =  $row["DEL"];
+
+		$articleNo   = htmlspecialchars($row["ARTICLENO"]);
+		$article     = htmlspecialchars($row["ARTICLE"]);
+		$room        = htmlspecialchars($row["ROOM"]);
+		$keyPlace    = htmlspecialchars($row["KEYPLACE"]);
+		$articleNote = htmlspecialchars($row["ARTICLENOTE"]);
+		$keyBox      = htmlspecialchars($row["KEYBOX"]);
+		$drawing     = htmlspecialchars($row["DRAWING"]);
+		$sellCharge  = htmlspecialchars($row["SELLCHARGE"]);
+		$del         = htmlspecialchars($row["DEL"]);
 
 		$purpose  = '更新';
 		$btnImage = 'btn_load.png';
@@ -187,7 +202,7 @@ function subArticleEdit()
 		$btnImage = 'btn_enter.png';
 	}
 
-	subMenu2();
+	subMenu();
 ?>
 	<script type="text/javascript" src="./js/article.js"></script>
 
@@ -212,17 +227,10 @@ function subArticleEdit()
 		<table border="0" cellpadding="5" cellspacing="1">
 			<tr>
 				<th>除外</th>
-				<?php if ($articleNo) { ?>
-					<td>
-						<input type="checkbox" name="del" value="1" checked="checked" /> 非除外
-						<input type="checkbox" name="del" value="0" <?php if ($del == '0') print ' checked="checked"' ?> /> 除外
-					</td>
-				<?php } else { ?>
-					<td>
-						<input type="radio" name="del" value="1" checked="checked" /> 非除外
-						<input type="radio" name="del" value="0" <?php if ($del == '0') print ' checked="checked"' ?> /> 除外
-					</td>
-				<?php } ?>
+				<td>
+					<input type="radio" name="del" value="1" checked="checked" /> 非除外
+					<input type="radio" name="del" value="0" <?php if ($del == '0') print ' checked="checked"' ?> /> 除外
+				</td>
 			</tr>
 			<tr>
 				<th>物件名<span class="red">（必須）</span></th>
@@ -262,7 +270,9 @@ function subArticleEdit()
 
 		<a href="javascript:form.act.value='articleSearch';form.submit();"><img src="./images/btn_return.png" /></a>
 
-		&nbsp;&nbsp;<a href="javascript:fnArticleDeleteCheck(<?php print $articleNo ?>);"><img src="./images/btn_del.png" /></a>
+		<?php if ($articleNo) { ?>
+			&nbsp;&nbsp;<a href="javascript:fnArticleDeleteCheck(<?php print $articleNo ?>);"><img src="./images/btn_del.png" /></a>
+		<?php } ?>
 	</form>
 <?php
 }
@@ -305,10 +315,13 @@ function subArticleEditComplete()
 	if ($articleNo) {
 		// 編集
 		$sql = fnSqlArticleUpdate($articleNo, $article, $room, $keyPlace, $address, $articleNote, $keyBox, $drawing, $sellCharge, $del);
+
 		$res = mysqli_query($conn, $sql);
 	} else {
 		// 新規登録
-		$sql = fnSqlArticleInsert(fnNextNo('ARTICLE'), $article, $room, $keyPlace, $address, $articleNote, $keyBox, $drawing, $sellCharge, $del);
+
+		$sql = fnSqlArticleInsert(fnNextNo('ARTICLE'), $article, $room, $keyPlace, $articleNote, $drawing, $address, $keyBox, $sellCharge, $del);
+
 
 		$res = mysqli_query($conn, $sql);
